+++ integration/src/bin/test-mmap.rs
@@ -7,7 +7,7 @@ fn main() {
   let p = dir.path().join("test_mmap");
   {
     const N: usize = 10;
-    let l = Arc::new(SkipMap::mmap_mut(1 << 20, &p, true).unwrap());
+    let l = Arc::new(SkipMap::mmap_mut(&p, 1 << 20, true).unwrap());
     for i in 0..N {
       let l = l.clone();
       std::thread::spawn(move || {
@@ -29,7 +29,7 @@ fn main() {
 
   {
     const N2: usize = 10;
-    let l = Arc::new(SkipMap::mmap(&p, false).unwrap());
+    let l = Arc::new(SkipMap::<u64>::mmap(&p, false).unwrap());
     assert_eq!(N2, l.len());
     for i in 0..N2 {
       let l = l.clone();