+++ src/map/iterator.rs
@@ -2,43 +2,6 @@ use core::ops::RangeFull;
 
 use super::*;
 
-/// A range over the skipmap. The current state of the iterator can be cloned by
-/// simply value copying the struct.
-pub struct MapRange<'a, T, C, Q: ?Sized = &'static str, R = RangeFull>(MapIterator<'a, T, C, Q, R>);
-
-impl<'a, T, C, Q, R> Clone for MapRange<'a, T, C, Q, R>
-where
-  R: Clone,
-  Q: Clone,
-  T: Clone,
-{
-  fn clone(&self) -> Self {
-    Self(self.0.clone())
-  }
-}
-
-impl<'a, T, C, Q, R> Copy for MapRange<'a, T, C, Q, R>
-where
-  R: Copy,
-  Q: Copy,
-  T: Copy,
-{
-}
-
-impl<'a, T, C, Q, R> core::ops::Deref for MapRange<'a, T, C, Q, R> {
-  type Target = MapIterator<'a, T, C, Q, R>;
-
-  fn deref(&self) -> &Self::Target {
-    &self.0
-  }
-}
-
-impl<'a, T, C, Q, R> core::ops::DerefMut for MapRange<'a, T, C, Q, R> {
-  fn deref_mut(&mut self) -> &mut Self::Target {
-    &mut self.0
-  }
-}
-
 /// An iterator over the skipmap. The current state of the iterator can be cloned by
 /// simply value copying the struct.
 pub struct MapIterator<'a, T, C, Q: ?Sized = &'static [u8], R = core::ops::RangeFull> {
@@ -47,7 +10,7 @@ pub struct MapIterator<'a, T, C, Q: ?Sized = &'static [u8], R = core::ops::Range
   pub(super) version: u64,
   pub(super) range: R,
   pub(super) all_versions: bool,
-  pub(super) last: Option<EntryRef<'a, T>>,
+  pub(super) last: Option<EntryRef<'a, T, C>>,
   pub(super) _phantom: core::marker::PhantomData<Q>,
 }
 
@@ -85,6 +48,25 @@ where
   }
 }
 
+impl<'a, Q, R, T, C> MapIterator<'a, T, C, Q, R>
+where
+  &'a [u8]: PartialOrd<Q>,
+  Q: ?Sized + PartialOrd<&'a [u8]>,
+{
+  #[inline]
+  pub(super) fn range(version: u64, map: &'a SkipMap<T, C>, r: R, all_versions: bool) -> Self {
+    Self {
+      map,
+      nd: map.head,
+      version,
+      range: r,
+      last: None,
+      all_versions,
+      _phantom: core::marker::PhantomData,
+    }
+  }
+}
+
 impl<'a, Q, R, T, C> MapIterator<'a, T, C, Q, R> {
   /// Returns the bounds of the iterator.
   #[inline]
@@ -101,94 +83,45 @@ where
   Q: ?Sized + PartialOrd<&'a [u8]>,
   R: RangeBounds<Q>,
 {
-  #[inline]
-  pub(super) fn range(
-    version: u64,
-    map: &'a SkipMap<T, C>,
-    r: R,
-    all_versions: bool,
-  ) -> MapRange<'a, T, C, Q, R> {
-    MapRange(Self {
-      map,
-      nd: map.head,
-      version,
-      range: r,
-      last: None,
-      all_versions,
-      _phantom: core::marker::PhantomData,
-    })
-  }
-
-  /// Seeks position at the first entry in map. Returns the key and value
-  /// if the iterator is pointing at a valid entry, and `None` otherwise.
-  pub fn first(&mut self) -> Option<EntryRef<T>> {
-    self.nd = self.map.first_in(self.version)?;
-
-    loop {
-      if self.nd.is_null() || self.nd.ptr == self.map.tail.ptr {
-        return None;
-      }
-
-      unsafe {
-        let node = self.nd.as_ptr();
-        let nk = node.get_key(&self.map.arena);
-
-        if node.trailer.version() > self.version {
-          self.nd = self.map.get_next(self.nd, 0);
-          continue;
-        }
-
-        if self.map.cmp.contains(&self.range, nk) {
-          let ent = EntryRef {
-            key: nk,
-            trailer: node.trailer,
-            value: node.get_value(&self.map.arena),
-          };
-          self.last = Some(ent);
-          return Some(ent);
-        }
-
-        self.nd = self.map.get_next(self.nd, 0);
-      }
+  /// Moves the iterator to the highest element whose key is below the given bound.
+  /// If no such element is found then `None` is returned.
+  pub fn seek_upper_bound(&mut self, upper: Bound<&[u8]>) -> Option<EntryRef<T, C>> {
+    match upper {
+      Bound::Included(key) => self.seek_le(key).map(|n| {
+        let ent = EntryRef::from_node(n, self.map);
+        self.last = Some(ent);
+        ent
+      }),
+      Bound::Excluded(key) => self.seek_lt(key).map(|n| {
+        let ent = EntryRef::from_node(n, self.map);
+        self.last = Some(ent);
+        ent
+      }),
+      Bound::Unbounded => self.last(),
     }
   }
 
-  /// Seeks position at the last entry in the iterator. Returns the key and value if
-  /// the iterator is pointing at a valid entry, and `None` otherwise.
-  pub fn last(&mut self) -> Option<EntryRef<T>> {
-    self.nd = self.map.last_in(self.version)?;
-
-    loop {
-      unsafe {
-        if self.nd.is_null() || self.nd.ptr == self.map.head.ptr {
-          return None;
-        }
-
-        let node = self.nd.as_ptr();
-        if node.trailer.version() > self.version {
-          self.nd = self.map.get_prev(self.nd, 0);
-          continue;
-        }
-
-        let nk = node.get_key(&self.map.arena);
-        if self.map.cmp.contains(&self.range, nk) {
-          let ent = EntryRef {
-            key: nk,
-            trailer: node.trailer,
-            value: node.get_value(&self.map.arena),
-          };
-          return Some(ent);
-        }
-
-        self.nd = self.map.get_prev(self.nd, 0);
-      }
+  /// Moves the iterator to the lowest element whose key is above the given bound.
+  /// If no such element is found then `None` is returned.
+  pub fn seek_lower_bound(&mut self, lower: Bound<&[u8]>) -> Option<EntryRef<T, C>> {
+    match lower {
+      Bound::Included(key) => self.seek_ge(key).map(|n| {
+        let ent = EntryRef::from_node(n, self.map);
+        self.last = Some(ent);
+        ent
+      }),
+      Bound::Excluded(key) => self.seek_gt(key).map(|n| {
+        let ent = EntryRef::from_node(n, self.map);
+        self.last = Some(ent);
+        ent
+      }),
+      Bound::Unbounded => self.first(),
     }
   }
 
   /// Advances to the next position. Returns the key and value if the
   /// iterator is pointing at a valid entry, and `None` otherwise.
-  #[allow(clippy::should_implement_trait)]
-  pub fn next(&mut self) -> Option<EntryRef<T>> {
+  fn next_in(&mut self) -> Option<EntryRef<T, C>> {
     loop {
       unsafe {
         self.nd = self.map.get_next(self.nd, 0);
@@ -214,6 +147,7 @@ where
 
         if self.map.cmp.contains(&self.range, nk) {
           let ent = EntryRef {
+            map: self.map,
             key: nk,
             trailer: node.trailer,
             value: node.get_value(&self.map.arena),
@@ -227,7 +161,7 @@ where
 
   /// Advances to the prev position. Returns the key and value if the
   /// iterator is pointing at a valid entry, and `None` otherwise.
-  pub fn prev(&mut self) -> Option<EntryRef<T>> {
+  fn prev(&mut self) -> Option<EntryRef<T, C>> {
     loop {
       unsafe {
         self.nd = self.map.get_prev(self.nd, 0);
@@ -253,6 +187,7 @@ where
 
         if self.map.cmp.contains(&self.range, nk) {
           let ent = EntryRef {
+            map: self.map,
             key: nk,
             trailer: node.trailer,
             value: node.get_value(&self.map.arena),
@@ -261,42 +196,6 @@ where
           return Some(ent);
         }
       }
-    }
-  }
-
-  /// Moves the iterator to the highest element whose key is below the given bound.
-  /// If no such element is found then `None` is returned.
-  pub fn seek_upper_bound(&mut self, upper: Bound<&[u8]>) -> Option<EntryRef<T>> {
-    match upper {
-      Bound::Included(key) => self.seek_le(key).map(|n| {
-        let ent = EntryRef::from_node(n, &self.map.arena);
-        self.last = Some(ent);
-        ent
-      }),
-      Bound::Excluded(key) => self.seek_lt(key).map(|n| {
-        let ent = EntryRef::from_node(n, &self.map.arena);
-        self.last = Some(ent);
-        ent
-      }),
-      Bound::Unbounded => self.last(),
-    }
-  }
-
-  /// Moves the iterator to the lowest element whose key is above the given bound.
-  /// If no such element is found then `None` is returned.
-  pub fn seek_lower_bound(&mut self, lower: Bound<&[u8]>) -> Option<EntryRef<T>> {
-    match lower {
-      Bound::Included(key) => self.seek_ge(key).map(|n| {
-        let ent = EntryRef::from_node(n, &self.map.arena);
-        self.last = Some(ent);
-        ent
-      }),
-      Bound::Excluded(key) => self.seek_gt(key).map(|n| {
-        let ent = EntryRef::from_node(n, &self.map.arena);
-        self.last = Some(ent);
-        ent
-      }),
-      Bound::Unbounded => self.first(),
     }
   }
 
@@ -456,5 +355,140 @@ where
         }
       }
     }
+  }
+
+  /// Seeks position at the first entry in map. Returns the key and value
+  /// if the iterator is pointing at a valid entry, and `None` otherwise.
+  fn first(&mut self) -> Option<EntryRef<T, C>> {
+    self.nd = self.map.first_in(self.version)?;
+
+    loop {
+      if self.nd.is_null() || self.nd.ptr == self.map.tail.ptr {
+        return None;
+      }
+
+      unsafe {
+        let node = self.nd.as_ptr();
+        let nk = node.get_key(&self.map.arena);
+
+        if node.trailer.version() > self.version {
+          self.nd = self.map.get_next(self.nd, 0);
+          continue;
+        }
+
+        if self.map.cmp.contains(&self.range, nk) {
+          let ent = EntryRef {
+            map: self.map,
+            key: nk,
+            trailer: node.trailer,
+            value: node.get_value(&self.map.arena),
+          };
+          self.last = Some(ent);
+          return Some(ent);
+        }
+
+        self.nd = self.map.get_next(self.nd, 0);
+      }
+    }
+  }
+
+  /// Seeks position at the last entry in the iterator. Returns the key and value if
+  /// the iterator is pointing at a valid entry, and `None` otherwise.
+  fn last(&mut self) -> Option<EntryRef<T, C>> {
+    self.nd = self.map.last_in(self.version)?;
+
+    loop {
+      unsafe {
+        if self.nd.is_null() || self.nd.ptr == self.map.head.ptr {
+          return None;
+        }
+
+        let node = self.nd.as_ptr();
+        if node.trailer.version() > self.version {
+          self.nd = self.map.get_prev(self.nd, 0);
+          continue;
+        }
+
+        let nk = node.get_key(&self.map.arena);
+        if self.map.cmp.contains(&self.range, nk) {
+          let ent = EntryRef {
+            map: self.map,
+            key: nk,
+            trailer: node.trailer,
+            value: node.get_value(&self.map.arena),
+          };
+          return Some(ent);
+        }
+
+        self.nd = self.map.get_prev(self.nd, 0);
+      }
+    }
+  }
+}
+
+impl<'a, Q, R, T, C> Iterator for MapIterator<'a, T, C, Q, R>
+where
+  C: Comparator,
+  T: Trailer,
+  &'a [u8]: PartialOrd<Q>,
+  Q: ?Sized + PartialOrd<&'a [u8]>,
+  R: RangeBounds<Q>,
+{
+  type Item = EntryRef<'a, T, C>;
+
+  #[inline]
+  fn next(&mut self) -> Option<Self::Item> {
+    self.next_in().map(|v| {
+      // Safety: the EntryRef holds a reference to the map, so it is always valid.
+      unsafe { core::mem::transmute(v) }
+    })
+  }
+
+  #[inline]
+  fn last(mut self) -> Option<Self::Item>
+  where
+    Self: Sized,
+  {
+    self.seek_upper_bound(Bound::Unbounded).map(|e| {
+      // Safety: the EntryRef holds a reference to the map, so it is always valid.
+      unsafe { core::mem::transmute(e) }
+    })
+  }
+
+  #[inline]
+  fn max(self) -> Option<Self::Item>
+  where
+    Self: Sized,
+    Self::Item: Ord,
+  {
+    self.last()
+  }
+
+  #[inline]
+  fn min(mut self) -> Option<Self::Item>
+  where
+    Self: Sized,
+    Self::Item: Ord,
+  {
+    self.first().map(|e| {
+      // Safety: the EntryRef holds a reference to the map, so it is always valid.
+      unsafe { core::mem::transmute(e) }
+    })
+  }
+}
+
+impl<'a, Q, R, T, C> DoubleEndedIterator for MapIterator<'a, T, C, Q, R>
+where
+  C: Comparator,
+  T: Trailer,
+  &'a [u8]: PartialOrd<Q>,
+  Q: ?Sized + PartialOrd<&'a [u8]>,
+  R: RangeBounds<Q>,
+{
+  fn next_back(&mut self) -> Option<Self::Item> {
+    self.prev().map(|v| {
+      // Safety: the EntryRef holds a reference to the map, so it is always valid.
+      unsafe { core::mem::transmute(v) }
+    })
   }
 }