+++ src/map/tests.rs
@@ -35,8 +35,8 @@ fn make_value(i: usize) -> std::vec::Vec<u8> {
 fn empty_in(l: SkipMap) {
   let mut it = l.iter_all_versions(0);
 
-  assert!(it.first().is_none());
-  assert!(it.last().is_none());
+  assert!(it.seek_lower_bound(Bound::Unbounded).is_none());
+  assert!(it.seek_upper_bound(Bound::Unbounded).is_none());
   assert!(it.seek_lower_bound(Bound::Included(b"aaa")).is_none());
   assert!(it.seek_upper_bound(Bound::Excluded(b"aaa")).is_none());
   assert!(it.seek_lower_bound(Bound::Excluded(b"aaa")).is_none());
@@ -51,6 +51,7 @@ fn empty_in(l: SkipMap) {
   assert!(!l.contains_key(0, b"aaa"));
   assert!(l.size() > 0);
   assert!(l.capacity() > 0);
+  assert_eq!(l.remaining(), l.capacity() - l.size());
 }
 
 #[test]
@@ -64,7 +65,7 @@ fn test_empty() {
 fn test_empty_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_empty_mmap_mut");
-  empty_in(SkipMap::mmap_mut(1000, p, true).unwrap());
+  empty_in(SkipMap::mmap_mut(p, 1000, true).unwrap());
 }
 
 #[test]
@@ -78,17 +79,21 @@ fn full_in(l: impl FnOnce(usize) -> SkipMap) {
   let l = l(1000);
   let mut found_arena_full = false;
 
+  let mut full_at = 0;
   for i in 0..100 {
     if let Err(e) = l.insert(0, &make_int_key(i), &make_value(i)) {
       assert!(matches!(e, Error::Full(_)));
       found_arena_full = true;
+      full_at = i;
       break;
     }
   }
 
   assert!(found_arena_full);
 
-  let e = l.insert(0, "someval".as_bytes(), &[]).unwrap_err();
+  let e = l
+    .insert(0, &make_int_key(full_at + 1), &make_value(full_at + 1))
+    .unwrap_err();
 
   assert!(matches!(e, Error::Full(_)));
 }
@@ -104,7 +109,7 @@ fn test_full() {
 fn test_full_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_full_mmap_mut");
-  full_in(|n| SkipMap::mmap_mut(n, p, true).unwrap());
+  full_in(|n| SkipMap::mmap_mut(p, n, true).unwrap());
 }
 
 #[test]
@@ -174,12 +179,27 @@ fn basic_in(mut l: SkipMap) {
 
   assert!(l.get_or_insert(2, b"c", &[]).unwrap().is_none());
 
+  {
+    #[allow(clippy::clone_on_copy)]
+    let a1 = l.get(1, b"a").unwrap().clone();
+    let a2 = l.get(2, b"a").unwrap();
+    assert!(a1 > a2);
+    assert_ne!(a1, a2);
+    let b1 = l.get(1, b"b").unwrap();
+    let b2 = l.get(2, b"b").unwrap();
+    assert!(b1 > b2);
+    assert_ne!(b1, b2);
+    let mut arr = [a1, a2, b1, b2];
+    arr.sort();
+    assert_eq!(arr, [a2, a1, b2, b1]);
+  }
+
   l.clear().unwrap();
 
   {
     let mut it = l.iter_all_versions(0);
-    assert!(it.first().is_none());
-    assert!(it.last().is_none());
+    assert!(it.seek_lower_bound(Bound::Unbounded).is_none());
+    assert!(it.seek_upper_bound(Bound::Unbounded).is_none());
   }
   assert!(l.is_empty());
 
@@ -201,7 +221,7 @@ fn test_basic() {
 fn test_basic_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_basic_mmap_mut");
-  basic_in(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  basic_in(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -246,38 +266,38 @@ fn iter_all_versions_mvcc(l: SkipMap) {
   assert_eq!(num, 4);
 
   let mut it = l.iter_all_versions(0);
-  assert!(it.first().is_none());
-  assert!(it.last().is_none());
+  assert!(it.seek_lower_bound(Bound::Unbounded).is_none());
+  assert!(it.seek_upper_bound(Bound::Unbounded).is_none());
 
   let mut it = l.iter_all_versions(1);
-  let ent = it.first().unwrap();
+  let ent = it.seek_lower_bound(Bound::Unbounded).unwrap();
   assert_eq!(ent.key(), b"a");
   assert_eq!(ent.value(), b"a1");
   assert_eq!(ent.trailer().version(), 1);
 
-  let ent = it.last().unwrap();
+  let ent = it.seek_upper_bound(Bound::Unbounded).unwrap();
   assert_eq!(ent.key(), b"c");
   assert_eq!(ent.value(), b"c1");
   assert_eq!(ent.trailer().version(), 1);
 
   let mut it = l.iter_all_versions(2);
-  let ent = it.first().unwrap();
+  let ent = it.seek_lower_bound(Bound::Unbounded).unwrap();
   assert_eq!(ent.key(), b"a");
   assert_eq!(ent.value(), b"a1");
   assert_eq!(ent.trailer().version(), 1);
 
-  let ent = it.last().unwrap();
+  let ent = it.seek_upper_bound(Bound::Unbounded).unwrap();
   assert_eq!(ent.key(), b"c");
   assert_eq!(ent.value(), b"c1");
   assert_eq!(ent.trailer().version(), 1);
 
   let mut it = l.iter_all_versions(3);
-  let ent = it.first().unwrap();
+  let ent = it.min().unwrap();
   assert_eq!(ent.key(), b"a");
   assert_eq!(ent.value(), b"a2");
   assert_eq!(ent.trailer().version(), 3);
 
-  let ent = it.last().unwrap();
+  let ent = it.max().unwrap();
   assert_eq!(ent.key(), b"c");
   assert_eq!(ent.value(), b"c2");
   assert_eq!(ent.trailer().version(), 3);
@@ -316,7 +336,7 @@ fn test_iter_all_versions_mvcc_mmap_mut() {
   let p = dir
     .path()
     .join("test_skipmap_iter_all_versions_mvcc_mmap_mut");
-  iter_all_versions_mvcc(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  iter_all_versions_mvcc(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -412,7 +432,7 @@ fn test_get_mvcc() {
 fn test_get_mvcc_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_get_mvcc_mmap_mut");
-  get_mvcc(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  get_mvcc(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -512,7 +532,7 @@ fn test_gt() {
 fn test_gt_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_gt_mmap_mut");
-  gt_in(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  gt_in(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -610,7 +630,7 @@ fn test_ge() {
 fn test_ge_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_ge_mmap_mut");
-  ge_in(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  ge_in(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -722,7 +742,7 @@ fn test_le() {
 fn test_le_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_le_mmap_mut");
-  gt_in(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  gt_in(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -816,7 +836,7 @@ fn test_lt() {
 fn test_lt_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_lt_mmap_mut");
-  lt_in(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  lt_in(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -858,7 +878,7 @@ fn test_basic_large_testcases_mmap_mut() {
   let p = dir
     .path()
     .join("test_skipmap_basic_large_testcases_mmap_mut");
-  let l = Arc::new(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  let l = Arc::new(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
   test_basic_large_testcases_in(l);
 }
 
@@ -911,7 +931,7 @@ fn test_concurrent_basic() {
 fn test_concurrent_basic_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_concurrent_basic_mmap_mut");
-  let l = Arc::new(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  let l = Arc::new(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
   test_concurrent_basic_runner(l);
 }
 
@@ -962,7 +982,7 @@ fn test_concurrent_basic_big_values_mmap_mut() {
   let p = dir
     .path()
     .join("test_skipmap_concurrent_basic_big_values_mmap_mut");
-  test_concurrent_basic_big_values_runner(Arc::new(SkipMap::mmap_mut(120 << 20, p, true).unwrap()));
+  test_concurrent_basic_big_values_runner(Arc::new(SkipMap::mmap_mut(p, 120 << 20, true).unwrap()));
 }
 
 #[test]
@@ -1031,7 +1051,7 @@ fn test_concurrent_one_key() {
 fn test_concurrent_one_key_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_concurrent_one_key_mmap_mut");
-  concurrent_one_key(Arc::new(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap()));
+  concurrent_one_key(Arc::new(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap()));
 }
 
 #[test]
@@ -1049,7 +1069,7 @@ fn iter_all_versionsator_next(l: SkipMap) {
   }
 
   let mut it = l.iter_all_versions(0);
-  let mut ent = it.first().unwrap();
+  let mut ent = it.seek_lower_bound(Bound::Unbounded).unwrap();
   for i in 0..N {
     assert_eq!(ent.key(), make_int_key(i));
     assert_eq!(ent.value(), make_value(i));
@@ -1074,7 +1094,7 @@ fn test_iter_all_versionsator_next_mmap_mut() {
   let p = dir
     .path()
     .join("test_skipmap_iter_all_versionsator_next_mmap_mut");
-  iter_all_versionsator_next(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  iter_all_versionsator_next(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -1093,7 +1113,7 @@ fn range_next(l: SkipMap) {
 
   let upper = make_int_key(50);
   let mut it = l.range(0, ..=upper.as_slice());
-  let mut ent = it.first();
+  let mut ent = it.seek_lower_bound(Bound::Unbounded);
   for i in 0..N {
     if i <= 50 {
       {
@@ -1122,7 +1142,7 @@ fn test_range_next() {
 fn test_range_next_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_range_next_mmap_mut");
-  iter_all_versionsator_next(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  iter_all_versionsator_next(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -1140,20 +1160,20 @@ fn iter_all_versionsator_prev(l: SkipMap) {
   }
 
   let mut it = l.iter_all_versions(0);
-  let mut ent = it.last().unwrap();
+  let mut ent = it.seek_upper_bound(Bound::Unbounded).unwrap();
   for i in (0..N).rev() {
     assert_eq!(ent.key(), make_int_key(i));
     assert_eq!(ent.value(), make_value(i));
     if i != 0 {
-      ent = it.prev().unwrap();
+      ent = it.next_back().unwrap();
     }
   }
 
-  assert!(it.prev().is_none());
+  assert!(it.next_back().is_none());
 }
 
 #[test]
-fn test_iter_all_versionsator_prev() {
+fn test_iter_all_versionsator_next_back() {
   iter_all_versionsator_prev(SkipMap::new(ARENA_SIZE).unwrap());
 }
 
@@ -1165,7 +1185,7 @@ fn test_iter_all_versionsator_prev_mmap_mut() {
   let p = dir
     .path()
     .join("test_skipmap_iter_all_versionsator_prev_mmap_mut");
-  iter_all_versionsator_prev(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  iter_all_versionsator_prev(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -1184,7 +1204,7 @@ fn range_prev(l: SkipMap) {
 
   let lower = make_int_key(50);
   let mut it = l.range(0, lower.as_slice()..);
-  let mut ent = it.last();
+  let mut ent = it.seek_upper_bound(Bound::Unbounded);
   for i in (0..N).rev() {
     if i >= 50 {
       {
@@ -1192,18 +1212,18 @@ fn range_prev(l: SkipMap) {
         assert_eq!(ent.key(), make_int_key(i));
         assert_eq!(ent.value(), make_value(i));
       }
-      ent = it.prev();
+      ent = it.next_back();
     } else {
       assert!(ent.is_none());
-      ent = it.prev();
+      ent = it.next_back();
     }
   }
 
-  assert!(it.prev().is_none());
+  assert!(it.next_back().is_none());
 }
 
 #[test]
-fn test_range_prev() {
+fn test_range_next_back() {
   range_prev(SkipMap::new(ARENA_SIZE).unwrap());
 }
 
@@ -1213,7 +1233,7 @@ fn test_range_prev() {
 fn test_range_prev_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_range_prev_mmap_mut");
-  range_prev(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  range_prev(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -1286,7 +1306,7 @@ fn test_iter_all_versionsator_seek_ge_mmap_mut() {
   let p = dir
     .path()
     .join("test_skipmap_iter_all_versionsator_seek_ge_mmap_mut");
-  iter_all_versionsator_seek_ge(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  iter_all_versionsator_seek_ge(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -1346,7 +1366,7 @@ fn test_iter_all_versionsator_seek_lt_mmap_mut() {
   let p = dir
     .path()
     .join("test_skipmap_iter_all_versionsator_seek_lt_mmap_mut");
-  iter_all_versionsator_seek_lt(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  iter_all_versionsator_seek_lt(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -1440,7 +1460,7 @@ fn range(l: SkipMap) {
   assert_eq!(ent.key(), make_int_key(3));
   assert_eq!(ent.value(), make_value(3));
 
-  assert!(it.prev().is_none());
+  assert!(it.next_back().is_none());
 }
 
 #[test]
@@ -1454,7 +1474,7 @@ fn test_range() {
 fn test_range_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_range_mmap_mut");
-  range(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  range(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -1504,7 +1524,7 @@ fn test_iter_latest() {
 fn test_iter_latest_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_iter_latest_mmap_mut");
-  iter_latest(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  iter_latest(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -1554,7 +1574,7 @@ fn test_range_latest() {
 fn test_range_latest_mmap_mut() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("test_skipmap_range_latest_mmap_mut");
-  range_latest(SkipMap::mmap_mut(ARENA_SIZE, p, true).unwrap());
+  range_latest(SkipMap::mmap_mut(p, ARENA_SIZE, true).unwrap());
 }
 
 #[test]
@@ -1571,7 +1591,7 @@ fn test_reopen_mmap() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("reopen_skipmap");
   {
-    let l = SkipMap::mmap_mut(ARENA_SIZE, &p, true).unwrap();
+    let l = SkipMap::mmap_mut(&p, ARENA_SIZE, true).unwrap();
     for i in 0..1000 {
       l.insert(0, &key(i), &new_value(i)).unwrap();
     }
@@ -1596,7 +1616,7 @@ fn test_reopen_mmap2() {
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("reopen_skipmap2");
   {
-    let l = SkipMap::mmap_mut_with_comparator(ARENA_SIZE, &p, true, Ascend).unwrap();
+    let l = SkipMap::mmap_mut_with_comparator(&p, ARENA_SIZE, true, Ascend).unwrap();
     for i in 0..1000 {
       l.insert(0, &key(i), &new_value(i)).unwrap();
     }